Unified SM2 and SM3 bytes API
